Add dry-run mode for copying files

+++ src/actions/file/copy.rs
@@ -44,7 +44,7 @@ impl Action for FileCopy {
         &self,
         manifest: &Manifest,
         context: &Context,
-        _dry_run: bool,
+        dry_run: bool,
     ) -> Result<ActionResult, ActionError> {
         let tera = self.init(manifest);
 
@@ -58,26 +58,32 @@ impl Action for FileCopy {
         let mut parent = PathBuf::from(&self.to);
         parent.pop();
 
-        debug!(
-            message = "Creating Prerequisite Directories",
-            directories = &parent.to_str().unwrap()
-        );
+        if !dry_run {
+            debug!(
+                message = "Creating Prerequisite Directories",
+                directories = &parent.to_str().unwrap()
+            );
 
-        create_dir_all(parent).context("Failed to create parent directory")?;
+            create_dir_all(parent).context("Failed to create parent directory")?;
 
-        let mut file = std::fs::File::create(self.to.clone()).context("Failed to create file")?;
+            let mut file =
+                std::fs::File::create(self.to.clone()).context("Failed to create file")?;
 
-        file.write_all(contents.as_bytes())
-            .context("Failed to create file")?;
+            file.write_all(contents.as_bytes())
+                .context("Failed to create file")?;
 
-        file.sync_all().context("Failed to create file")?;
+            file.sync_all().context("Failed to create file")?;
 
-        set_permissions(PathBuf::from(self.to.clone()), self.chmod)
-            .context("Failed to set permissions")?;
-
-        Ok(ActionResult {
-            message: String::from("Copied"),
-        })
+            set_permissions(PathBuf::from(self.to.clone()), self.chmod)
+                .context("Failed to set permissions")?;
+            Ok(ActionResult {
+                message: String::from("Copied"),
+            })
+        } else {
+            Ok(ActionResult {
+                message: format!("copy from {} to {}", self.from, self.to),
+            })
+        }
     }
 }
